Added typing. Resolved feedback from Mathis. Made corrections to the tutorial notebook

+++ src/lava/lib/optimization/solvers/qp/models.py
@@ -7,6 +7,7 @@ Implement behaviors (models) of the processes defined in processes.py
 For further documentation please refer to processes.py
 """
 import numpy as np
+import typing as ty
 from lava.magma.core.sync.protocols.loihi_protocol import LoihiProtocol
 from lava.magma.core.model.py.ports import PyInPort, PyOutPort
 from lava.magma.core.model.py.type import LavaPyType
@@ -28,11 +29,9 @@ from lava.lib.optimization.solvers.qp.processes import (
 @implements(proc=ConstraintDirections, protocol=LoihiProtocol)
 @requires(CPU)
 class PyCDModel(PyLoihiProcessModel):
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    weights: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, float)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, float)
+    weights: np.ndarray = LavaPyType(np.ndarray, float)
 
     def run_spk(self):
         s_in = self.s_in.recv()
@@ -45,11 +44,9 @@ class PyCDModel(PyLoihiProcessModel):
 @implements(proc=ConstraintNeurons, protocol=LoihiProtocol)
 @requires(CPU)
 class PyCNeuModel(PyLoihiProcessModel):
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    thresholds: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
+    thresholds: np.ndarray = LavaPyType(np.ndarray, np.float64)
 
     def run_spk(self):
         s_in = self.s_in.recv()
@@ -62,11 +59,9 @@ class PyCNeuModel(PyLoihiProcessModel):
 @implements(proc=QuadraticConnectivity, protocol=LoihiProtocol)
 @requires(CPU)
 class PyQCModel(PyLoihiProcessModel):
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    weights: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
+    weights: np.ndarray = LavaPyType(np.ndarray, np.float64)
 
     def run_spk(self):
         s_in = self.s_in.recv()
@@ -79,28 +74,18 @@ class PyQCModel(PyLoihiProcessModel):
 @implements(proc=SolutionNeurons, protocol=LoihiProtocol)
 @requires(CPU)
 class PySNModel(PyLoihiProcessModel):
-    s_in_qc: PyInPort = LavaPyType(
-        PyInPort.VEC_DENSE, np.float64, precision=24
-    )
-    a_out_qc: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    s_in_cn: PyInPort = LavaPyType(
-        PyInPort.VEC_DENSE, np.float64, precision=24
-    )
-    a_out_cc: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    qp_neuron_state: np.ndarray = LavaPyType(
-        np.ndarray, np.float64, precision=24
-    )
-    grad_bias: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    alpha: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    beta: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    alpha_decay_schedule: int = LavaPyType(int, np.int32, precision=24)
-    beta_growth_schedule: int = LavaPyType(int, np.int32, precision=24)
-    decay_counter: int = LavaPyType(int, np.int32, precision=24)
-    growth_counter: int = LavaPyType(int, np.int32, precision=24)
+    s_in_qc: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    a_out_qc: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
+    s_in_cn: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    a_out_cc: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
+    qp_neuron_state: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    grad_bias: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    alpha: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    beta: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    alpha_decay_schedule: int = LavaPyType(int, np.int32)
+    beta_growth_schedule: int = LavaPyType(int, np.int32)
+    decay_counter: int = LavaPyType(int, np.int32)
+    growth_counter: int = LavaPyType(int, np.int32)
 
     def run_spk(self):
         a_out = self.qp_neuron_state
@@ -133,11 +118,9 @@ class PySNModel(PyLoihiProcessModel):
 @implements(proc=ConstraintNormals, protocol=LoihiProtocol)
 @requires(CPU)
 class PyCNorModel(PyLoihiProcessModel):
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
-    weights: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
+    weights: np.ndarray = LavaPyType(np.ndarray, np.float64)
 
     def run_spk(self):
         s_in = self.s_in.recv()
@@ -151,16 +134,10 @@ class PyCNorModel(PyLoihiProcessModel):
 class SubCCModel(AbstractSubProcessModel):
     """Implement constraintCheckProcess behavior via sub Processes."""
 
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    constraint_matrix: np.ndarray = LavaPyType(
-        np.ndarray, np.float64, precision=24
-    )
-    constraint_bias: np.ndarray = LavaPyType(
-        np.ndarray, np.float64, precision=24
-    )
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    constraint_matrix: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    constraint_bias: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
 
     def __init__(self, proc):
         """Builds sub Process structure of the Process."""
@@ -194,22 +171,19 @@ class SubCCModel(AbstractSubProcessModel):
 class SubGDModel(AbstractSubProcessModel):
     """Implement gradientDynamics Process behavior via sub Processes."""
 
-    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64, precision=24)
-    hessian: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
+    s_in: PyInPort = LavaPyType(PyInPort.VEC_DENSE, np.float64)
+    hessian: np.ndarray = LavaPyType(np.ndarray, np.float64)
     constraint_matrix_T: np.ndarray = LavaPyType(
-        np.ndarray, np.float64, precision=24
+        np.ndarray,
+        np.float64,
     )
-    grad_bias: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    qp_neuron_state: np.ndarray = LavaPyType(
-        np.ndarray, np.float64, precision=24
-    )
-    alpha: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    beta: np.ndarray = LavaPyType(np.ndarray, np.float64, precision=24)
-    alpha_decay_schedule: int = LavaPyType(int, np.int32, precision=24)
-    beta_growth_schedule: int = LavaPyType(int, np.int32, precision=24)
-    a_out: PyOutPort = LavaPyType(
-        PyOutPort.VEC_DENSE, np.float64, precision=24
-    )
+    grad_bias: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    qp_neuron_state: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    alpha: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    beta: np.ndarray = LavaPyType(np.ndarray, np.float64)
+    alpha_decay_schedule: int = LavaPyType(int, np.int32)
+    beta_growth_schedule: int = LavaPyType(int, np.int32)
+    a_out: PyOutPort = LavaPyType(PyOutPort.VEC_DENSE, np.float64)
 
     def __init__(self, proc):
         """Builds sub Process structure of the Process."""
